Created Dockerfile and updated dir App. Small fix in api.py bcz main.py moved in to dir App.
