+++ Dockerfile
@@ -1,0 +1,7 @@
+FROM python:3.11
+WORKDIR /app
+COPY ./requirements.txt ./requirements.txt
+RUN pip install --no-cache-dir --upgrade -r requirements.txt
+COPY ./ ./
+EXPOSE 80
+CMD ["uvicorn", "app.api:app", "--host", "0.0.0.0", "--port", "80"]
